[WIP] Moved some files around and organized project
