+++ FlightTracking/Services_Utilities/LocationUtilities.swift
@@ -39,15 +39,4 @@ struct LocationUtilities {
 
     return formattedDistance
   }
-
 }
-extension Double {
-  var degreesToRadians: Double {
-    return self * .pi / 180.0
-  }
-
-  var radiansToDegrees: Double {
-    return self * 180.0 / .pi
-  }
-}
-
