+++ FlightTracking/Views/FlightDetailViews/GateDepartureBanner.swift
