+++ FlightTracking/Views/FlightDetailViews/InfoSection/InfoSectionViewModel.swift
